bapl1_l1_ex_adding-more-operators
a) Add a remainder operator ('%') to the language, with the same priority as the other multiplicative operators.
b) Add an exponential operator ('^') to the language, with a higher priority than the multiplicative operators. Use the same concepts we used when we added multiplicative operators. You will probably need a new kind of expression, besides 'term' and 'exp'.