bapl1_l2_ex_floating-point-numbers
a) Change the definition of numeral so that it can match floating-point numbers like "0.1"
and "13.25". It is up to you whether the syntax would also accept things like "0." or ".5".
(You probably won't need to change anything in the captures for this feature as all operations we used naturally handle floats)
b) Add scientific notation ("E notation") to the previous definition, so that we can write numbers like "0.5E-3" and "123e12".
(Again, you probably won't need to change anything in the captures)