bapl1_l5_ex_logical-operators
- Implement the logical operators 'and' and 'or', with short
circuit. "Short circuit" means that the second operand is evaluated
only when needed.  For 'and', the second operand is evaluated only when
the first results in true.  For 'or', the second operand is evaluated
only when the first results in false. As is usual in dynamic languages,
the result of the operation should be the last result that was computed.
For instance, "4 and 5" results in 5, "0 and 3" results in 0,
"0 or 10" results in 10, and "2 or 3" results in 2.
(Hint: for the 'and', create a new instruction "jmpZP", that
checks whether the value on the top of the stack is zero;
if so, it jumps to a given address without popping the top value;
otherwise, it pops the top value and continues without jumping.
For the ;5D'or', create a dual instruction "jmpNZP", which is similar
to jmpZP but with the test inverted.)