bapl1_l8_ex_checking-redeclaration-of-variables
Add a check to the compiler so that it raises an error if we declare a local variable in a point where another local variable with the same name is in scope.
Correct the previous check so that it allows shadowing of variables in inner blocks. Two variables with the same name in the same block should raise the error, but there should be no errors if they are in different blocks (even if one is nested into the other).